* Editores base concluidos

+++ src/com/maton/tools/stiletto/model/Frame.java
@@ -2,11 +2,11 @@ package com.maton.tools.stiletto.model;
 
 public class Frame {
 	private Sprite source;
-	private float time;
+	private int time;
 	
 	public Frame(Sprite source) {
 		this.source = source;
-		time = 100.0f;
+		time = 100;
 	}
 
 	public Sprite getSource() {
@@ -17,11 +17,11 @@ public class Frame {
 		this.source = source;
 	}
 
-	public float getTime() {
+	public int getTime() {
 		return time;
 	}
 
-	public void setTime(float time) {
+	public void setTime(int time) {
 		this.time = time;
 	}
 }
